Файл main-model разбит на 3 файла
