+++ src/main.js
@@ -1,7 +1,12 @@
 import Presenter from './presenter/main-presenter';
-import MainModel from './model/main-model';
+import PointsModel from './model/points-model';
+import OffersModel from './model/offers-model';
+import DestinationsModel from './model/destinations-model';
 
-const mainModel = new MainModel();
-const presenter = new Presenter({mainModel});
+const pointsModel = new PointsModel();
+const offersModel = new OffersModel();
+const destinationsModel = new DestinationsModel();
+
+const presenter = new Presenter({pointsModel, offersModel, destinationsModel});
 
 presenter.init();
