+++ src/model/destinations-model.js
@@ -1,0 +1,7 @@
+import { destinationsMock } from '../mock/destinations-mock';
+
+export default class DestinationsModel {
+  get destinations() {
+    return [...destinationsMock];
+  }
+}
