+++ src/model/offers-model.js
@@ -1,0 +1,7 @@
+import { offersMock } from '../mock/offers-mock';
+
+export default class OffersModel {
+  get offers() {
+    return [...offersMock];
+  }
+}
