+++ src/model/points-model.js
@@ -1,0 +1,7 @@
+import { pointsMock } from '../mock/points-mock';
+
+export default class PointsModel {
+  get points() {
+    return [...pointsMock];
+  }
+}
